Refactoring

+++ Assets/Scripts/MathUtil.cs
@@ -2,8 +2,8 @@ using System;
 
 public class MathUtil
 {
-    public static double GeometricDistribution(int x, float p)
+    public static float GeometricDistribution(int x, float p)
     {
-        return Math.Pow(1 - p, x - 1) * p;
+        return (float)Math.Pow(1 - p, x - 1) * p;
     }
 }
